+++ trabalho/bolhaParada.h
@@ -1,7 +1,7 @@
 #ifndef BOLHAPARADA_H_INCLUDED
 #define BOLHAPARADA_H_INCLUDED
 
-void ordenarBolhaParada(int n, int vet[], int *cont, int *troca);
+void ordenarBolhaParada(int n, int vet[], long long int *cont, long long int *troca);
 
 
 #endif // BOLHAPARADA_H_INCLUDED
