+++ trabalho/insercaoBinaria.h
@@ -1,7 +1,7 @@
 #ifndef INSERCAOBINARIA_H_INCLUDED
 #define INSERCAOBINARIA_H_INCLUDED
 
-void ordenarInsercaoBinaria(int n, int vet[], int *contador, int *trocas);
+void ordenarInsercaoBinaria(int n, int vet[], long long int *contador, long long int *trocas);
