+++ trabalho/insercaoDireta.h
@@ -1,7 +1,7 @@
 #ifndef INSERCAODIRETA_H_INCLUDED
 #define INSERCAODIRETA_H_INCLUDED
 
-void ordenarInsercaoDireta(int n, int vet[],int *contador, int *trocas);
+void ordenarInsercaoDireta(int n, int vet[],long long int *contador, long long int *trocas);
 
 
 #endif // INSERCAODIRETA_H_INCLUDED
