+++ trabalho/mergesort.h
@@ -1,6 +1,6 @@
 #ifndef MERGESORT_H_INCLUDED
 #define MERGESORT_H_INCLUDED
 
-void ordenarMergeSort(int *vetor, int tamanho, int *contador, int *trocas);
+void ordenarMergeSort(int *vetor, int tamanho, long long int *contador, long long int *trocas);
 
 #endif // MERGESORT_H_INCLUDED
