+++ trabalho/quicksortIni.h
@@ -1,7 +1,6 @@
 #ifndef QUICKSORT_H_INCLUDED
 #define QUICKSORT_H_INCLUDED
 
-void ordenarQuickSortIni(int *vetor, int tamanho,int *contador,int *trocas);
-//void ordenarQuickSortCentro(int vet[], int esquerda, int direita);
+void ordenarQuickSortIni(int *vetor, int tamanho,long long int *contador,long long int *trocas);
 
 #endif // QUICKSORT_H_INCLUDED
