+++ trabalho/quicksortMediana.h
@@ -1,6 +1,6 @@
 #ifndef QUICKSORTMEDIANA_H_INCLUDED
 #define QUICKSORTMEDIANA_H_INCLUDED
 
-void ordenarQuickSortMediana(int* vetor,int tamanho,int *contador, int *trocas);
+void ordenarQuickSortMediana(int* vetor,int tamanho,long long int *contador, long long int *trocas);
 
 #endif // QUICKSORTMEDIANA_H_INCLUDED
