+++ trabalho/radixsort.h
@@ -1,6 +1,6 @@
 #ifndef RADIXSORT_H_INCLUDED
 #define RADIXSORT_H_INCLUDED
 
-void ordenarRadixSort(int *vetor,int tamanho,int *contador,int *trocas);
+void ordenarRadixSort(int *vetor,int tamanho,long long int *contador,long long int *trocas);
 
 #endif // RADIXSORT_H_INCLUDED
