+++ trabalho/selecaoDireta.h
@@ -1,7 +1,7 @@
 #ifndef SELECAODIRETA_H_INCLUDED
 #define SELECAODIRETA_H_INCLUDED
 
-void ordenarSelecaoDireta(int n, int vet[], int *contador, int *trocas);
+void ordenarSelecaoDireta(int n, int vet[], long long int *contador, long long int *trocas);
