+++ trabalho/shellsort.c
@@ -2,11 +2,13 @@
 #include <stdlib.h>
 #include "shellsort.h"
 
-void ordenarShellSort(int tamanho, int *vetor, int *comparacoes, int *trocas){
+void ordenarShellSort(int tamanho, int *vetor, long long int *comparacoes, long long int *trocas){
     int aux=0,gap=tamanho/2,i=0,j=0,cont=1;
 
     //Ordenação
     while(gap>0){
+
+        //Ordena os valores de um sub-vetor
         for(int x=gap;x<tamanho;x++){
             j = x;
             i = j - gap;
@@ -25,6 +27,8 @@ void ordenarShellSort(int tamanho, int *vetor, int *comparacoes, int *trocas){
                 i = j - gap;
             }
         }
+
+        //Reduz o tamanho do espaçamento entre os elementos
         gap = gap/2;
     }
 }
