+++ trabalho/shellsort.h
@@ -1,7 +1,7 @@
 #ifndef SHELLSORT_H_INCLUDED
 #define SHELLSORT_H_INCLUDED
 
-void ordenarShellSort(int n, int vet[], int *contador, int *trocas);
+void ordenarShellSort(int n, int vet[], long long int *contador, long long int *trocas);
 
 
 #endif // SHELLSORT_H_INCLUDED
